backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
python-dotenv
langgraph
langchain-google-genai
firestore
sqlmodel
psycopg2-binary
agentscope
websockets
pinecone-client
httpx
pydantic[email]
google-api-python-client
google-auth-oauthlib
notion-client
slack-sdk
PyGithub
python-jose[cryptography]
passlib[bcrypt]
